pyproject.toml
[tool.poetry]
name = "xss-detection-gpt4"
version = "0.1.0"
description = ""
authors = ["dripchar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
